+++ packages/storefront/src/app/scss/border/examples/border.stories.ts
@@ -12,18 +12,48 @@ export const borderRadiusScssStory: Story<'div'> = {
       children: [
         {
           tag: 'div',
-          properties: { className: 'rounded-sm border p-fluid-sm' },
-          children: ['border-radius: $pds-border-radius-small'],
+          properties: { className: 'rounded-xs border p-fluid-md' },
+          children: ['border-radius: $border-radius-xs;'],
         },
         {
           tag: 'div',
-          properties: { className: 'rounded-md border p-fluid-sm' },
-          children: ['border-radius: $pds-border-radius-medium'],
+          properties: { className: 'rounded-sm border p-fluid-md' },
+          children: ['border-radius: $border-radius-sm;'],
         },
         {
           tag: 'div',
-          properties: { className: 'rounded-lg border p-fluid-sm' },
-          children: ['border-radius: $pds-border-radius-large'],
+          properties: { className: 'rounded-md border p-fluid-md' },
+          children: ['border-radius: $border-radius-md;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-lg border p-fluid-md' },
+          children: ['border-radius: $border-radius-lg;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-xl border p-fluid-md' },
+          children: ['border-radius: $border-radius-xl;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-2xl border p-fluid-md' },
+          children: ['border-radius: $border-radius-2xl;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-3xl border p-fluid-md' },
+          children: ['border-radius: $border-radius-3xl;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-4xl border p-fluid-md' },
+          children: ['border-radius: $border-radius-4xl;'],
+        },
+        {
+          tag: 'div',
+          properties: { className: 'rounded-full border p-fluid-md' },
+          children: ['border-radius: $border-radius-full;'],
         },
       ],
     },
@@ -41,16 +71,16 @@ export const borderWidthScssStory: Story<'div'> = {
         {
           tag: 'div',
           properties: {
-            className: 'border-thin p-fluid-sm',
+            className: 'border-1 p-fluid-md',
           },
-          children: ['border-width: $pds-border-width-thin'],
+          children: ['border-width: $border-width-1;'],
         },
         {
           tag: 'div',
           properties: {
-            className: 'border-regular p-fluid-sm',
+            className: 'border-2 p-fluid-md',
           },
-          children: ['border-width: $pds-border-width-base'],
+          children: ['border-width: $border-width-2;'],
         },
       ],
     },