+++ packages/styles/projects/scss/src/motion.ts
@@ -10,12 +10,20 @@ import {
 
 export const getMotionScss = () => {
   return `
-    $pds-motion-duration-long: ${motionDurationLong};
-    $pds-motion-duration-moderate: ${motionDurationModerate};
-    $pds-motion-duration-short: ${motionDurationShort};
-    $pds-motion-duration-very-long: ${motionDurationVeryLong};
-    $pds-motion-easing-base: ${motionEaseInOut};
-    $pds-motion-easing-in: ${motionEaseIn};
-    $pds-motion-easing-out: ${motionEaseOut};
+    $motion-duration-sm: ${motionDurationShort};
+    $motion-duration-md: ${motionDurationModerate};
+    $motion-duration-lg: ${motionDurationLong};
+    $motion-duration-xl: ${motionDurationVeryLong};
+    $motion-ease-in-out: ${motionEaseInOut};
+    $motion-ease-in: ${motionEaseIn};
+    $motion-ease-out: ${motionEaseOut};
+
+    $pds-motion-duration-long: ${motionDurationLong}; /* alias (deprecated) */
+    $pds-motion-duration-moderate: ${motionDurationModerate}; /* alias (deprecated) */
+    $pds-motion-duration-short: ${motionDurationShort}; /* alias (deprecated) */
+    $pds-motion-duration-very-long: ${motionDurationVeryLong}; /* alias (deprecated) */
+    $pds-motion-easing-base: ${motionEaseInOut}; /* alias (deprecated) */
+    $pds-motion-easing-in: ${motionEaseIn}; /* alias (deprecated) */
+    $pds-motion-easing-out: ${motionEaseOut}; /* alias (deprecated) */
 `;
 };