+++ packages/styles/src/app/scss/ScssBorder.tsx
@@ -2,13 +2,19 @@ export const ScssBorder = () => {
   return (
     <div className="scss-border">
       <div className="scss-border__radius-wrapper">
+        <div className="scss-border__radius scss-border__radius--xs">Border Radius X-Small</div>
         <div className="scss-border__radius scss-border__radius--sm">Border Radius Small</div>
         <div className="scss-border__radius scss-border__radius--md">Border Radius Medium</div>
         <div className="scss-border__radius scss-border__radius--lg">Border Radius Large</div>
+        <div className="scss-border__radius scss-border__radius--xl">Border Radius X-Large</div>
+        <div className="scss-border__radius scss-border__radius--2xl">Border Radius 2X-Large</div>
+        <div className="scss-border__radius scss-border__radius--3xl">Border Radius 3X-Large</div>
+        <div className="scss-border__radius scss-border__radius--4xl">Border Radius 4X-Large</div>
+        <div className="scss-border__radius scss-border__radius--full">Border Radius Full</div>
       </div>
       <div className="scss-border__width-wrapper">
-        <div className="scss-border__width scss-border__width--thin">Border Width Thin</div>
-        <div className="scss-border__width scss-border__width--regular">Border Width Regular</div>
+        <div className="scss-border__width scss-border__width--1">Border Width 1</div>
+        <div className="scss-border__width scss-border__width--2">Border Width 2</div>
       </div>
     </div>
   );